+++ packages/sdk-ts/src/client/indexer/rest/index.ts
@@ -1,3 +1,4 @@
 export { IndexerRestDerivativesChronosApi } from './IndexerRestDerivativesChronosApi'
 export { IndexerRestSpotChronosApi } from './IndexerRestSpotChronosApi'
 export { IndexerRestExplorerApi } from './IndexerRestExplorerApi'
+export { IndexerRestMarketChronosApi } from './IndexerRestMarketChronosApi'