+++ packages/sdk-ts/src/core/exchange/msgs/MsgCreateBinaryOptionsLimitOrder.ts
@@ -5,7 +5,10 @@ import {
   OrderTypeMap,
 } from '@injectivelabs/chain-api/injective/exchange/v1beta1/exchange_pb'
 import { MsgBase } from '../../MsgBase'
-import { amountToCosmosSdkDecAmount } from '../../../utils/numbers'
+import {
+  amountToCosmosSdkDecAmount,
+  getTriggerPrice,
+} from '../../../utils/numbers'
 
 export declare namespace MsgCreateBinaryOptionsLimitOrder {
   export interface Params {
@@ -84,9 +87,7 @@ export default class MsgCreateBinaryOptionsLimitOrder extends MsgBase<
       ...initialParams,
       price: amountToCosmosSdkDecAmount(initialParams.price).toFixed(),
       margin: amountToCosmosSdkDecAmount(initialParams.margin).toFixed(),
-      triggerPrice: amountToCosmosSdkDecAmount(
-        initialParams.triggerPrice || 0,
-      ).toFixed(),
+      triggerPrice: getTriggerPrice(initialParams.triggerPrice),
       quantity: amountToCosmosSdkDecAmount(initialParams.quantity).toFixed(),
     } as MsgCreateBinaryOptionsLimitOrder.Params