+++ packages/sdk-ts/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.109](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.108...@injectivelabs/sdk-ts@1.0.109) (2022-08-29)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
+## [1.0.108](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.107...@injectivelabs/sdk-ts@1.0.108) (2022-08-26)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
+## [1.0.107](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.106...@injectivelabs/sdk-ts@1.0.107) (2022-08-26)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
+## [1.0.106](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.105...@injectivelabs/sdk-ts@1.0.106) (2022-08-26)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
+## [1.0.105](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.104...@injectivelabs/sdk-ts@1.0.105) (2022-08-26)
+
+
+### Bug Fixes
+
+* trigger price ([9a9294b](https://github.com/InjectiveLabs/injective-ts/commit/9a9294b46ea3b40125ae597ef3b1f07ab893412a))
+
+
+
+
+
+## [1.0.104](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.103...@injectivelabs/sdk-ts@1.0.104) (2022-08-26)
+
+
+### Bug Fixes
+
+* snakecase fields ([96b5b47](https://github.com/InjectiveLabs/injective-ts/commit/96b5b4763f9b672e1b8e9b7782fce36671e0edd7))
+
+
+### Features
+
+* order history ([#65](https://github.com/InjectiveLabs/injective-ts/issues/65)) ([c179735](https://github.com/InjectiveLabs/injective-ts/commit/c179735d5b3f7168a017073720721332e44319cb))
+
+
+
+
+
 ## [1.0.103](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.0.102...@injectivelabs/sdk-ts@1.0.103) (2022-08-25)
 
 **Note:** Version bump only for package @injectivelabs/sdk-ts